Bug Hunter/requirements.txt
openai
langchain
selenium
stable-baselines3
requests
beautifulsoup4
python-dotenv
pytest 